+++ core/src/auto_enum/args.rs
@@ -1,9 +1,6 @@
 use proc_macro2::{token_stream::IntoIter, Delimiter, Group, Ident, TokenStream, TokenTree};
 use quote::ToTokens;
-use smallvec::smallvec;
 use syn::{Path, Result};
-
-use crate::utils::Stack;
 
 // =============================================================================
 // Arg
@@ -68,29 +65,27 @@ impl Eq for Arg {}
 // =============================================================================
 // Parse
 
-macro_rules! arg_err {
+macro_rules! error {
     ($msg:expr) => {
-        syn::Error::new($msg.span(), format!("invalid arguments: {}", $msg))
+        syn::Error::new($msg.span(), $msg)
     };
     ($span:expr, $msg:expr) => {
-        Err(syn::Error::new($span.span(), format!("invalid arguments: {}", $msg)))
+        return Err(syn::Error::new($span.span(), $msg))
     };
     ($span:expr, $($tt:tt)*) => {
-        arg_err!($span, format!($($tt)*))
+        error!($span, format!($($tt)*))
     };
 }
 
-pub(super) fn parse_group(group: TokenStream) -> Result<(Stack<Arg>, Option<String>, bool)> {
-    syn::parse2::<Group>(group)
-        .map_err(|e| arg_err!(e))
-        .and_then(|group| parse_args(group.stream()))
+pub(super) fn parse_group(group: TokenStream) -> Result<(Vec<Arg>, Option<String>, bool)> {
+    syn::parse2::<Group>(group).map_err(|e| error!(e)).and_then(|group| parse_args(group.stream()))
 }
 
-pub(super) fn parse_args(args: TokenStream) -> Result<(Stack<Arg>, Option<String>, bool)> {
+pub(super) fn parse_args(args: TokenStream) -> Result<(Vec<Arg>, Option<String>, bool)> {
     const ERR: &str = "expected one of `,`, `::`, or identifier, found ";
 
     let mut iter = args.into_iter();
-    let mut args = Stack::new();
+    let mut args = Vec::new();
     let mut marker = None;
     let mut never = false;
     while let Some(tt) = iter.next() {
@@ -108,7 +103,7 @@ pub(super) fn parse_args(args: TokenStream) -> Result<(Stack<Arg>, Option<String
                     }
 
                     if never {
-                        arg_err!(i, "multiple `never` option")?;
+                        error!(i, "multiple `never` option")
                     }
                     never = true;
                 }
@@ -116,17 +111,17 @@ pub(super) fn parse_args(args: TokenStream) -> Result<(Stack<Arg>, Option<String
             },
             TokenTree::Punct(p) => match p.as_char() {
                 ',' => {}
-                ':' => args.push(parse_path(smallvec![p.into()], &mut iter)?),
-                _ => arg_err!(p, "{}`{}`", ERR, p)?,
+                ':' => args.push(parse_path(vec![p.into()], &mut iter)?),
+                _ => error!(p, "{}`{}`", ERR, p),
             },
-            _ => arg_err!(tt, "{}`{}`", ERR, tt)?,
+            _ => error!(tt, "{}`{}`", ERR, tt),
         }
     }
 
     Ok((args, marker, never))
 }
 
-fn parse_path(mut path: Stack<TokenTree>, iter: &mut IntoIter) -> Result<Arg> {
+fn parse_path(mut path: Vec<TokenTree>, iter: &mut IntoIter) -> Result<Arg> {
     for tt in iter {
         match tt {
             TokenTree::Punct(ref p) if p.as_char() == ',' => break,
@@ -134,7 +129,7 @@ fn parse_path(mut path: Stack<TokenTree>, iter: &mut IntoIter) -> Result<Arg> {
         }
     }
 
-    syn::parse2(path.into_iter().collect()).map_err(|e| arg_err!(e)).map(Arg::Path)
+    syn::parse2(path.into_iter().collect()).map_err(|e| error!(e)).map(Arg::Path)
 }
 
 fn path_or_ident(ident: Ident, tt: Option<TokenTree>, iter: &mut IntoIter) -> Result<Arg> {
@@ -144,56 +139,56 @@ fn path_or_ident(ident: Ident, tt: Option<TokenTree>, iter: &mut IntoIter) -> Re
         None => Ok(Arg::Ident(ident)),
         Some(TokenTree::Punct(p)) => match p.as_char() {
             ',' => Ok(Arg::Ident(ident)),
-            ':' => parse_path(smallvec![ident.into(), p.into()], iter),
-            _ => arg_err!(p, "{}`{}`", ERR, p),
+            ':' => parse_path(vec![ident.into(), p.into()], iter),
+            _ => error!(p, "{}`{}`", ERR, p),
         },
-        Some(tt) => arg_err!(tt, "{}`{}`", ERR, tt),
+        Some(tt) => error!(tt, "{}`{}`", ERR, tt),
     }
 }
 
 fn marker_opt(
     ident: Ident,
     iter: &mut IntoIter,
-    args: &mut Stack<Arg>,
+    args: &mut Vec<Arg>,
     marker: &mut Option<String>,
 ) -> Result<()> {
     match iter.next() {
         Some(TokenTree::Group(ref g)) if g.delimiter() != Delimiter::Parenthesis => {
-            arg_err!(g, "invalid delimiter")?
+            error!(g, "invalid delimiter")
         }
         Some(TokenTree::Group(ref g)) => {
             let mut g = g.stream().into_iter();
             match g.next() {
                 Some(TokenTree::Ident(ref i)) if marker.is_some() => {
-                    arg_err!(i, "multiple `marker` option")?
+                    error!(i, "multiple `marker` option")
                 }
                 Some(TokenTree::Ident(i)) => *marker = Some(i.to_string()),
-                Some(tt) => arg_err!(tt, "expected an identifier, found `{}`", tt)?,
-                None => arg_err!(ident, "empty `marker` option")?,
+                Some(tt) => error!(tt, "expected an identifier, found `{}`", tt),
+                None => error!(ident, "empty `marker` option"),
             }
             match g.next() {
                 None => {}
                 Some(TokenTree::Punct(ref p)) if p.as_char() == ',' => {
                     if let Some(tt) = g.next() {
-                        arg_err!(tt, "multiple identifier in `marker` option")?;
+                        error!(tt, "multiple identifier in `marker` option")
                     }
                 }
-                Some(tt) => arg_err!(tt, "multiple identifier in `marker` option")?,
+                Some(tt) => error!(tt, "multiple identifier in `marker` option"),
             }
         }
         Some(TokenTree::Punct(ref p)) if p.as_char() == '=' => {
             match iter.next() {
                 Some(TokenTree::Ident(ref i)) if marker.is_some() => {
-                    arg_err!(i, "multiple `marker` option")?
+                    error!(i, "multiple `marker` option")
                 }
                 Some(TokenTree::Ident(i)) => *marker = Some(i.to_string()),
-                Some(tt) => arg_err!(tt, "expected an identifier, found `{}`", tt)?,
-                None => arg_err!(p, "empty `marker` option")?,
+                Some(tt) => error!(tt, "expected an identifier, found `{}`", tt),
+                None => error!(p, "empty `marker` option"),
             }
             match iter.next() {
                 None => {}
                 Some(TokenTree::Punct(ref p)) if p.as_char() == ',' => {}
-                Some(tt) => arg_err!(tt, "multiple identifier in `marker` option")?,
+                Some(tt) => error!(tt, "multiple identifier in `marker` option"),
             }
         }
         tt => args.push(path_or_ident(ident, tt, iter)?),