+++ core/src/utils/mod.rs
@@ -1,32 +1,17 @@
-use std::mem;
+use std::{iter, mem};
 
 use proc_macro2::{Ident, Span, TokenStream};
-use smallvec::SmallVec;
 use syn::{
-    punctuated::Punctuated, Block, Expr, ExprBlock, ExprTuple, ExprVerbatim, Path, PathSegment,
-    Stmt,
+    punctuated::Punctuated, token, Attribute, Block, Expr, ExprBlock, ExprCall, ExprPath,
+    ExprTuple, ExprVerbatim, Path, PathSegment, Stmt,
 };
 
-pub(crate) type Stack<T> = SmallVec<[T; 4]>;
+mod attrs;
+
+pub(crate) use self::attrs::{Attrs, AttrsMut};
 
 // =============================================================================
 // Extension traits
-
-pub(crate) trait OptionExt {
-    fn replace_boxed_expr<F: FnOnce(Expr) -> Expr>(&mut self, f: F);
-}
-
-impl OptionExt for Option<Box<Expr>> {
-    fn replace_boxed_expr<F: FnOnce(Expr) -> Expr>(&mut self, f: F) {
-        if self.is_none() {
-            self.replace(Box::new(unit()));
-        }
-
-        if let Some(expr) = self {
-            replace_expr(&mut **expr, f);
-        }
-    }
-}
 
 pub(crate) trait VecExt<T> {
     fn find_remove<P: FnMut(&T) -> bool>(&mut self, predicate: P) -> Option<T>;
@@ -41,10 +26,6 @@ impl<T> VecExt<T> for Vec<T> {
 // =============================================================================
 // Functions
 
-pub(crate) fn default<T: Default>() -> T {
-    T::default()
-}
-
 pub(crate) fn ident<S: AsRef<str>>(s: S) -> Ident {
     Ident::new(s.as_ref(), Span::call_site())
 }
@@ -54,15 +35,32 @@ pub(crate) fn path<I: IntoIterator<Item = PathSegment>>(segments: I) -> Path {
 }
 
 pub(crate) fn block(stmts: Vec<Stmt>) -> Block {
-    Block { brace_token: default(), stmts }
+    Block { brace_token: token::Brace::default(), stmts }
 }
 
 pub(crate) fn expr_block(block: Block) -> Expr {
     Expr::Block(ExprBlock { attrs: Vec::new(), label: None, block })
 }
 
+pub(crate) fn expr_call(attrs: Vec<Attribute>, path: Path, arg: Expr) -> Expr {
+    Expr::Call(ExprCall {
+        attrs,
+        func: Box::new(Expr::Path(ExprPath { attrs: Vec::new(), qself: None, path })),
+        paren_token: token::Paren::default(),
+        args: iter::once(arg).collect(),
+    })
+}
+
+pub(super) fn expr_compile_error(e: &syn::Error) -> Expr {
+    syn::parse2(e.to_compile_error()).unwrap()
+}
+
 pub(crate) fn unit() -> Expr {
-    Expr::Tuple(ExprTuple { attrs: Vec::new(), paren_token: default(), elems: Punctuated::new() })
+    Expr::Tuple(ExprTuple {
+        attrs: Vec::new(),
+        paren_token: token::Paren::default(),
+        elems: Punctuated::new(),
+    })
 }
 
 pub(crate) fn replace_expr<F: FnOnce(Expr) -> Expr>(this: &mut Expr, op: F) {
@@ -82,20 +80,18 @@ macro_rules! span {
     };
 }
 
-macro_rules! err {
+macro_rules! error {
+    // FIXME: syntax
+    (span => $span:expr, $msg:expr) => {
+        syn::Error::new($span, $msg)
+    };
     ($msg:expr) => {
         syn::Error::new_spanned(span!($msg), $msg)
-    };
-    ($span:ident .span(), $msg:expr) => {
-        syn::Error::new_spanned($span.span(), $msg)
     };
     ($span:expr, $msg:expr) => {
         syn::Error::new_spanned(span!($span), $msg)
     };
-    ($span:ident .span(), $($tt:tt)*) => {
-        err!($span.span(), format!($($tt)*))
-    };
     ($span:expr, $($tt:tt)*) => {
-        err!($span, format!($($tt)*))
+        error!($span, format!($($tt)*))
     };
 }