+++ derive/src/derive/core/iter/exact_size_iterator.rs
@@ -2,7 +2,7 @@ use crate::utils::*;
 
 pub(crate) const NAME: &[&str] = &["ExactSizeIterator"];
 
-pub(crate) fn derive(data: &Data, stack: &mut Stack<ItemImpl>) -> Result<()> {
+pub(crate) fn derive(data: &Data, items: &mut Vec<ItemImpl>) -> Result<()> {
     #[cfg(not(feature = "exact_size_is_empty"))]
     let is_empty = quote!();
     #[cfg(feature = "exact_size_is_empty")]
@@ -23,5 +23,5 @@ pub(crate) fn derive(data: &Data, stack: &mut Stack<ItemImpl>) -> Result<()> {
             }
         }?,
     )
-    .map(|item| stack.push(item))
+    .map(|item| items.push(item))
 }