+++ test_suite/compile-test/tests/ui/enum_derive/variant.rs
@@ -1,0 +1,37 @@
+// compile-fail
+
+#![deny(warnings)]
+
+use auto_enums::enum_derive;
+
+#[enum_derive(Clone)]
+enum Enum2<B> {
+    A,
+    B(B),
+}
+
+#[enum_derive(Clone)]
+enum Enum3 {
+    A = 2,
+    B,
+}
+
+#[enum_derive(Clone)]
+enum Enum4<A, B> {
+    A { x: A },
+    B(B),
+}
+
+#[enum_derive(Clone)]
+enum Enum5<B> {
+    A(),
+    B(B),
+}
+
+#[enum_derive(Clone)]
+enum Enum6<A> {
+    A(A),
+    B(A, B),
+}
+
+fn main() {}