+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+* Renamed `#[rec]` to `#[nested]`.
+
+* Improved error messages.
+
 # 0.5.8 - 2019-05-22
 
 * Added support for `futures::io::{AsyncSeek, AsyncBufRead}`.